requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
pandas
psycopg2-binary